requirements.txt
streamlit
python-dotenv
youtube_transcript_api
gtts
numpy
google.generativeai